distribution.R to decimal

+++ Mod/lab_2/db/migrate/20150319131121_create_distributions.rb
@@ -1,7 +1,7 @@
 class CreateDistributions < ActiveRecord::Migration
   def change
     create_table :distributions do |t|
-      t.integer :R
+      t.decimal :R
       t.integer :a
       t.integer :b
       t.integer :mx
